requirements.txt
asn1crypto==0.24.0
astroid==2.3.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
configparser==4.0.2
cryptography==2.7
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.4.1
pyodbc==4.0.27
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.25.6
wrapt==1.11.2
